populating database

+++ app/Http/Controllers/AdminController.php
@@ -5,7 +5,7 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Auth;
-
+use Illuminate\Support\Facades\Gate;
 use App\Models\Poll;
 use App\Models\Attendee;
 use App\Models\ChooseOption;
@@ -23,6 +23,7 @@ class AdminController extends Controller
      */
     public function showUsers()
     {
+      Gate::authorize('admin', Auth::user());
       $users = User::where('is_admin', '!=', true)->paginate(5);
       return view('pages.adminUsers',['users'=>$users]);
     }
@@ -34,6 +35,7 @@ class AdminController extends Controller
      */
     public function showEvents()
     {
+      Gate::authorize('admin', Auth::user());
       $events = Event::paginate(5);
       return view('pages.adminEvents',['events'=>$events]);
     }
@@ -45,6 +47,7 @@ class AdminController extends Controller
      */
     public function showReports()
     {
+      Gate::authorize('admin', Auth::user());
       $reports = Report::paginate(5);
       return view('pages.adminReports',['reports'=>$reports]);
     }
@@ -55,6 +58,10 @@ class AdminController extends Controller
      * @return Redirect back to the page
      */
     public function deleteUser($id){
+
+        Gate::authorize('admin', Auth::user());
+
+
         $user = User::where(['id'=>$id]);
         $user->delete();
         return redirect()->back();
@@ -66,6 +73,9 @@ class AdminController extends Controller
      * @return Redirect back to the page
      */
       public function blockUser($id){
+
+        Gate::authorize('admin', Auth::user());
+
         DB::table('users')->where(['id'=>$id])->update(['is_blocked'=>TRUE]);
         return redirect()->back();
       }
@@ -76,6 +86,8 @@ class AdminController extends Controller
      * @return Redirect back to the page
      */
     public function unblockUser($id){
+      Gate::authorize('admin', Auth::user());
+
       DB::table('users')->where(['id'=>$id])->update(['is_blocked'=>FALSE]);
       return redirect()->back();
     }
@@ -86,6 +98,9 @@ class AdminController extends Controller
      * @return Redirect back to the page
      */
     public function deleteEvent($id){
+
+      Gate::authorize('admin', Auth::user());
+
       $event = Event::where(['id'=>$id]);
       $event->delete();
       return redirect()->back();
